backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]>=2.0.40
aiosqlite==0.20.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
authlib==1.4.0
httpx==0.28.1
pydantic-settings==2.7.0
python-multipart==0.0.19
email-validator>=2.0.0
